refactor: Extension 파일분리
