+++ Calculator/Calculator/Model/Extension.swift
@@ -7,13 +7,3 @@
 
 import Foundation
 
-extension Operator: CalculatorItem { }
-
-extension Double: CalculatorItem { }
-
-extension String {
-    func split(with target: Character) -> [String] {
-        
-        return split(separator: target).map { String($0) }
-    }
-}
