+++ Calculator/Calculator/Model/Extension/Double+Extension.swift
@@ -1,0 +1,10 @@
+//
+//  Double_Extension.swift
+//  Calculator
+//
+//  Created by kimseongjun on 2023/01/31.
+//
+
+import Foundation
+
+extension Double: CalculatorItem { }
