+++ Calculator/Calculator/Model/Extension/Operator+Extension.swift
@@ -1,0 +1,10 @@
+//
+//  Operator+Extension.swift
+//  Calculator
+//
+//  Created by kimseongjun on 2023/01/31.
+//
+
+import Foundation
+
+extension Operator: CalculatorItem { }
